remove CustomChannel, adding getChannel from Call object
