+++ src/php/composer.json
@@ -5,10 +5,12 @@
   "version": "1.13.0",
   "require": {
     "php": ">=5.5.0",
-    "google/protobuf": "^v3.3.0"
+    "google/protobuf": "^v3.3.0",
+    "google/auth": "^1.3",
+    "google/gax": "^0.33.0",
+    "google/cloud": "^0.67.0"
   },
   "require-dev": {
-    "google/auth": "v0.9"
   },
   "autoload": {
     "psr-4": {
