--- src/php/lib/Grpc/CustomChannel.php
@@ -1,34 +1,0 @@
-<?php
-/*
- *
- * Copyright 2018 gRPC authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-namespace Grpc;
-
-// Since there is no private class for PHP.
-// TODO: put this class defination under Internal directory and add comment "Internel only"
-
-/**
- * CustomChannel can be used to customize and manage channels.
- * It can intercept the call object created by this channel.
- */
-interface CustomChannel
-{
-    // This function should be override
-    public function _UnaryStreamCallFactory();
-    public function _UnaryUnaryCallFactory($deserialize);
-}
