+++ src/php/lib/Grpc/Internal/InterceptorChannel.php
@@ -35,7 +35,8 @@ class InterceptorChannel extends \Grpc\Channel
     public function __construct($channel, $interceptor)
     {
         if (!is_a($channel, 'Grpc\Channel') &&
-            !is_a($channel, 'Grpc\Internal\InterceptorChannel')) {
+            !is_a($channel, 'Grpc\Internal\InterceptorChannel') &&
+            !is_a($channel, 'Grpc\GrpcExtensionChannel')) {
             throw new \Exception('The channel argument is not a Channel object '.
                 'or an InterceptorChannel object created by '.
                 'Interceptor::intercept($channel, Interceptor|Interceptor[] $interceptors)');
